requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
beanie==1.26.0
motor==3.4.0
# Motor/Beanie dependen de APIs internas de PyMongo; versiones nuevas han removido `_QUERY_OPTIONS`.
# Pin para evitar `ImportError: cannot import name '_QUERY_OPTIONS' from 'pymongo.cursor'`.
pymongo==4.6.3

# HTTP client
httpx==0.27.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# passlib 1.7.4 is not compatible with bcrypt 5.x (breaks backend detection and
# can trigger the 72-byte bcrypt password limit even for short passwords).
bcrypt<5

# Config
pydantic-settings==2.2.1
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
respx==0.21.1
mongomock-motor==0.0.21
